@@ -2,35 +2,37 @@
   "name": "example-app",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "license": "MIT",
   "scripts": {
     "build": "tsc",
-    "start": "yarn build && node build/index.js",
-    "clean": "rm -rf build && mkdir build",
-    "dev": "yarn clean && concurrently \"wait-on build/src/index.js && nodemon --ext '.js' --watch ../build --watch ./build --ignore 'cypress/**/*.js' node build/src/index.js\" \"yarn build --watch\"",
+    "start": "yarn build && node dist/example-app/src/index.js",
+    "clean": "rm -rf dist && mkdir dist",
+    "dev": "yarn clean && concurrently \"wait-on dist/src/index.js && nodemon --ext '.js' --watch ../build --watch ./dist --ignore 'cypress/**/*.js' node dist/example-app/src/index.js\" \"yarn build --watch\"",
     "cypress:open": "cypress open",
     "cypress:run": "cypress run"
   },
   "devDependencies": {
-    "@types/express": "^4.17.7",
-    "concurrently": "^5.2.0",
-    "cypress": "^4.11.0",
+    "@types/express": "^4.17.17",
+    "concurrently": "^8.0.1",
     "nodemon": "^2.0.4",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7",
-    "wait-on": "^5.2.0"
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.2",
+    "wait-on": "^7.0.1"
   },
   "dependencies": {
-    "@adminjs/express": "^3.0.0",
-    "@adminjs/typeorm": "1.4.0",
-    "@adminjs/upload": "../",
+    "@adminjs/design-system": "^4.0.0-beta-v4.8",
+    "@adminjs/express": "^6.0.0-beta.3",
+    "@adminjs/typeorm": "5.0.0-beta.1",
     "@google-cloud/storage": "^5.3.0",
-    "adminjs": "^3.3.1",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
+    "adminjs": "7.0.0-beta-v7.4",
+    "cypress": "^12.9.0",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
     "express-formidable": "^1.2.0",
-    "express-session": "^1.17.1",
+    "express-session": "^1.17.3",
     "pg": "^8.3.0",
-    "typeorm": "^0.2.25"
-  }
+    "typeorm": "^0.3.12"
+  },
+  "relativeDependencies": {}
 }
