@@ -1,17 +1,22 @@
 {
   "name": "@adminjs/upload",
-  "version": "2.0.1",
-  "main": "index.js",
-  "types": "types/index.d.ts",
+  "version": "4.0.2",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": "./build/index.js",
+      "types": "./types/index.d.ts"
+    }
+  },
   "private": false,
   "repository": "git@github.com:SoftwareBrothers/adminjs-upload.git",
   "license": "MIT",
   "scripts": {
     "release": "semantic-release",
-    "build": "yarn tsc",
+    "build": "tsc",
     "clean": "rm -fr build && mkdir build && rm -fr types && mkdir types",
     "dev": "yarn clean && tsc --watch",
-    "test": "TS_NODE_FILES=true mocha -r ts-node/register ./src/**/*.spec.ts",
+    "test": "TS_NODE_FILES=true mocha --loader=ts-node/esm ./src/**/*.spec.ts",
     "lint": "eslint './src/**/*' './example-app/**/*'",
     "check:all": "yarn lint && yarn build && yarn test"
   },
@@ -21,39 +26,44 @@
     }
   },
   "peerDependencies": {
-    "adminjs": ">=5.0.0"
+    "adminjs": "^7.0.0"
   },
   "optionalDependencies": {
-    "@google-cloud/storage": "^5.3.0",
-    "aws-sdk": "^2.728.0"
+    "@aws-sdk/client-s3": "^3.301.0",
+    "@aws-sdk/s3-request-presigner": "^3.301.0",
+    "@google-cloud/storage": "^6.9.4"
   },
   "devDependencies": {
-    "@commitlint/cli": "^8.3.5",
-    "@commitlint/config-conventional": "^8.3.4",
-    "@google-cloud/storage": "^5.3.0",
-    "@semantic-release/git": "^9.0.0",
-    "@types/chai": "^4.2.12",
-    "@types/mocha": "^8.0.3",
-    "@types/sinon": "^9.0.5",
-    "@types/sinon-chai": "^3.2.4",
-    "@typescript-eslint/eslint-plugin": "^3.7.0",
-    "@typescript-eslint/parser": "^3.7.0",
-    "adminjs": "^5.0.0",
-    "aws-sdk": "^2.728.0",
-    "chai": "^4.2.0",
-    "eslint": "^7.5.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-react-hooks": "^4.0.8",
+    "@aws-sdk/client-s3": "^3.301.0",
+    "@aws-sdk/s3-request-presigner": "^3.301.0",
+    "@commitlint/cli": "^17.4.4",
+    "@commitlint/config-conventional": "^17.4.4",
+    "@google-cloud/storage": "^6.9.4",
+    "@semantic-release/git": "^10.0.1",
+    "@types/chai": "^4.3.4",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.15.11",
+    "@types/sinon": "^10.0.13",
+    "@types/sinon-chai": "^3.2.9",
+    "@typescript-eslint/eslint-plugin": "^5.56.0",
+    "@typescript-eslint/parser": "^5.56.0",
+    "adminjs": "^7.0.0",
+    "aws-sdk-js-codemod": "^0.14.1",
+    "chai": "^4.3.7",
+    "eslint": "^8.36.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^4.2.5",
-    "mocha": "^8.1.3",
-    "semantic-release": "^17.0.7",
-    "semantic-release-slack-bot": "^1.6.2",
-    "sinon": "^9.0.3",
-    "sinon-chai": "^3.5.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^3.9.7"
-  }
+    "mocha": "^10.2.0",
+    "semantic-release": "^20.1.3",
+    "semantic-release-slack-bot": "^4.0.0",
+    "sinon": "^15.0.2",
+    "sinon-chai": "^3.7.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.2"
+  },
+  "dependencies": {}
 }
