@@ -1,9 +1,10 @@
 /* eslint-disable no-console */
 /* eslint-disable class-methods-use-this */
-import uploadFeature, { BaseProvider } from '@adminjs/upload'
+import uploadFeature, { BaseProvider } from '../../../../../src/index.js'
+import { componentLoader } from '../../component-loader.js'
 
-import { CreateResourceResult } from '../create-resource-result.type'
-import { Custom } from '../../../custom/custom.entity'
+import { CreateResourceResult } from '../create-resource-result.type.js'
+import { Custom } from '../../../custom/custom.entity.js'
 
 class MyProvider extends BaseProvider {
   constructor() {
@@ -31,6 +32,7 @@ const createPhotoResource = (): CreateResourceResult<typeof Custom> => ({
     properties: { file: { isVisible: { list: true, show: true, edit: true, filter: false } } },
   },
   features: [uploadFeature({
+    componentLoader,
     provider: new MyProvider(),
     properties: { file: 'file', key: 'filePath', bucket: 'bucket', filePath: 'path' },
     validation: { mimeTypes: ['image/png'] },
