@@ -1,7 +1,7 @@
-import uploadFeature from '@adminjs/upload'
-
-import { CreateResourceResult } from '../create-resource-result.type'
-import { Post } from '../../../post/post.entity'
+import uploadFeature from '../../../../../src/index.js'
+import { componentLoader } from '../../component-loader.js'
+import { CreateResourceResult } from '../create-resource-result.type.js'
+import { Post } from '../../../post/post.entity.js'
 
 const createPostResource = (): CreateResourceResult<typeof Post> => ({
   resource: Post,
@@ -9,9 +9,11 @@ const createPostResource = (): CreateResourceResult<typeof Post> => ({
     listProperties: ['id', 'bucketKey', 'bucket', 'path'],
   },
   features: [uploadFeature({
+    componentLoader,
     provider: {
       gcp: {
-        bucket: process.env.GCP_STORAGE_BUCKET,
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        bucket: process.env.GCP_STORAGE_BUCKET!,
         expires: 0,
       },
     },
