@@ -1,7 +1,8 @@
-import uploadFeature from '@adminjs/upload'
-import { CreateResourceResult } from '../create-resource-result.type'
-import credentials from '../../../credentials'
-import { User } from '../../../user/user.entity'
+import uploadFeature from '../../../../../src/index.js'
+import { componentLoader } from '../../component-loader.js'
+import { CreateResourceResult } from '../create-resource-result.type.js'
+import credentials from '../../../credentials.js'
+import { User } from '../../../user/user.entity.js'
 
 const createUserResource = (): CreateResourceResult<typeof User> => ({
   resource: User,
@@ -9,6 +10,7 @@ const createUserResource = (): CreateResourceResult<typeof User> => ({
     listProperties: ['id', 'avatar', 'email', 'test'],
   },
   features: [uploadFeature({
+    componentLoader,
     provider: { aws: credentials },
     properties: {
       filename: 'name',
