@@ -14,26 +14,26 @@ import {
 @Entity({ name: 'posts' })
 export class Post extends BaseEntity {
   @PrimaryGeneratedColumn()
-  public id: number;
+  public id: number
 
   @Column({ nullable: true })
-  public name: string;
+  public name: string
 
   @Column({ nullable: true, name: 'key' })
-  public bucketKey: string;
+  public bucketKey: string
 
   @Column({ nullable: true })
-  public bucket: string;
+  public bucket: string
 
   @Column({ nullable: true })
-  public path: string;
+  public path: string
 
   @Column({ nullable: true })
-  public mime: string;
+  public mime: string
 
   @CreateDateColumn({ name: 'created_at' })
-  public createdAt: Date;
+  public createdAt: Date
 
   @UpdateDateColumn({ name: 'updated_at' })
-  public updatedAt: Date;
+  public updatedAt: Date
 }
