@@ -14,32 +14,32 @@ import {
 @Entity({ name: 'users' })
 export class User extends BaseEntity {
   @PrimaryGeneratedColumn()
-  public id: number;
+  public id: number
 
   @Column({ nullable: true })
-  public name: string;
+  public name: string
 
   @Column({ nullable: true })
-  public fullname: string;
+  public fullname: string
 
   @Column()
-  public email: string;
+  public email: string
 
   @Column({ nullable: true })
-  public sajz: number;
+  public sajz: number
 
   @Column({ nullable: true })
-  public majmtajp: string;
+  public majmtajp: string
 
   @Column({ nullable: true, name: 's3_bucket' })
-  public s3Bucket: string;
+  public s3Bucket: string
 
   @Column({ nullable: true })
-  public avatar: string;
+  public avatar: string
 
   @CreateDateColumn({ name: 'created_at' })
-  public createdAt: Date;
+  public createdAt: Date
 
   @UpdateDateColumn({ name: 'updated_at' })
-  public updatedAt: Date;
+  public updatedAt: Date
 }
