@@ -1,7 +1,7 @@
-import React, { FC } from 'react'
 import { ShowPropertyProps } from 'adminjs'
+import React, { FC } from 'react'
 
-import File from './file'
+import File from './file.js'
 
 const List: FC<ShowPropertyProps> = (props) => (<File width={100} {...props} />)
 
