@@ -1,7 +1,12 @@
-import { BulkActionResponse, After, ActionRequest, ActionContext } from 'adminjs/types/src'
-import { BaseProvider } from '../providers'
-import { UploadOptionsWithDefault } from '../types/upload-options.type'
-import { deleteFile } from '../utils/delete-file'
+import {
+  ActionContext,
+  ActionRequest,
+  After,
+  BulkActionResponse,
+} from 'adminjs'
+import { BaseProvider } from '../providers/index.js'
+import { UploadOptionsWithDefault } from '../types/upload-options.type.js'
+import { deleteFile } from '../utils/delete-file.js'
 
 export const deleteFilesFactory = (
   uploadOptionsWithDefault: UploadOptionsWithDefault,
@@ -12,10 +17,13 @@ export const deleteFilesFactory = (
   context: ActionContext,
 ): Promise<BulkActionResponse> {
   const { records = [] } = context
-
-  await Promise.all(records.map(async (record) => {
-    await deleteFile(uploadOptionsWithDefault, provider, context, record)
-  }))
+  if (request.method === 'post') {
+    await Promise.all(
+      records.map(async (record) => {
+        await deleteFile(uploadOptionsWithDefault, provider, context, record)
+      }),
+    )
+  }
 
   return response
 }
