@@ -1,9 +1,11 @@
 /* eslint-disable no-underscore-dangle */
-import { ActionContext, ActionRequest, Before } from 'adminjs'
+import { ActionContext, ActionRequest, Before, ComponentLoader } from 'adminjs'
 import { expect } from 'chai'
-import { CONTEXT_NAMESPACE } from '../constants'
-import UploadOptions, { UploadOptionsWithDefault } from '../types/upload-options.type'
-import { stripPayloadFactory } from './strip-payload-factory'
+import { CONTEXT_NAMESPACE } from '../constants.js'
+import UploadOptions, { UploadOptionsWithDefault } from '../types/upload-options.type.js'
+import { stripPayloadFactory } from './strip-payload-factory.js'
+
+const componentLoader = new ComponentLoader()
 
 describe('stripPayloadFactory', () => {
   let actionContext = {} as ActionContext
@@ -13,6 +15,7 @@ describe('stripPayloadFactory', () => {
 
   beforeEach(() => {
     uploadOptions = {
+      componentLoader,
       properties: {
         key: 's3Key',
         filePath: 'resolvedPath',
@@ -21,7 +24,8 @@ describe('stripPayloadFactory', () => {
       },
       provider: {
         aws: { bucket: 'any' },
-      } as UploadOptions['provider'] }
+      } as UploadOptions['provider'],
+    }
   })
 
   context('standard upload payload', () => {
