@@ -1,13 +1,13 @@
-import { After, RecordActionResponse, ActionContext, BaseRecord, ActionRequest, UploadedFile, BaseResource } from 'adminjs'
-import sinon, { createStubInstance, SinonStubbedInstance } from 'sinon'
+import { ActionContext, ActionRequest, After, BaseRecord, BaseResource, ComponentLoader, RecordActionResponse, UploadedFile } from 'adminjs'
 import { expect } from 'chai'
+import sinon, { SinonStubbedInstance, createStubInstance } from 'sinon'
+import { CONTEXT_NAMESPACE } from '../constants.js'
+import { BaseProvider } from '../providers/index.js'
+import stubProvider from '../spec/stub-provider.js'
+import UploadOptions, { UploadOptionsWithDefault } from '../types/upload-options.type.js'
+import { updateRecordFactory } from './update-record-factory.js'
 
-import { updateRecordFactory } from './update-record-factory'
-import stubProvider from '../spec/stub-provider'
-import { BaseProvider } from '../providers'
-import UploadOptions, { UploadOptionsWithDefault } from '../types/upload-options.type'
-import { CONTEXT_NAMESPACE } from '../constants'
-
+const componentLoader = new ComponentLoader()
 describe('updateRecordFactory', () => {
   const request: ActionRequest = { method: 'post' } as ActionRequest
   let response: RecordActionResponse
@@ -29,10 +29,15 @@ describe('updateRecordFactory', () => {
 
   beforeEach(() => {
     provider = stubProvider(resolvedS3Path)
-    response = { record: { params: {
-      name: 'some value',
-    } } } as unknown as RecordActionResponse
+    response = {
+      record: {
+        params: {
+          name: 'some value',
+        },
+      },
+    } as unknown as RecordActionResponse
     uploadOptions = {
+      componentLoader,
       properties: {
         key: 's3Key',
         filePath: 'resolvedPath',
@@ -41,7 +46,8 @@ describe('updateRecordFactory', () => {
       },
       provider: {
         aws: { bucket: 'any' },
-      } as UploadOptions['provider'] }
+      } as UploadOptions['provider'],
+    }
     recordStub = createStubInstance(BaseRecord, {
       id: sinon.stub<any, string>().returns('1'),
       isValid: sinon.stub<any, boolean>().returns(true),
