@@ -1,5 +1,5 @@
-export * from './aws-provider'
-export * from './base-provider'
-export * from './gcp-provider'
-export * from './local-provider'
-export * from './minio-provider'
+export * from './aws-provider.js'
+export * from './base-provider.js'
+export * from './gcp-provider.js'
+export * from './local-provider.js'
+export * from './minio-provider.js'
