@@ -1,14 +1,15 @@
+import AdminJS, { ActionContext, ActionRequest, After, ComponentLoader, RecordActionResponse } from 'adminjs'
 import chai, { expect } from 'chai'
 import sinon from 'sinon'
 import sinonChai from 'sinon-chai'
-import { After, RecordActionResponse, ActionRequest, ActionContext } from 'adminjs'
-import { BaseProvider } from './providers/base-provider'
-import UploadOptions from './types/upload-options.type'
-
-import uploadFile from './upload-file.feature'
-import stubProvider from './spec/stub-provider'
+import { BaseProvider } from './providers/base-provider.js'
+import stubProvider from './spec/stub-provider.js'
+import UploadOptions from './types/upload-options.type.js'
+import uploadFile from './upload-file.feature.js'
 
 chai.use(sinonChai)
+
+const componentLoader = new ComponentLoader()
 
 describe('uploadFileFeature', () => {
   let provider: BaseProvider
@@ -43,15 +44,20 @@ describe('uploadFileFeature', () => {
     const key = 'someKeyValue'
 
     const getAfterHook = (options: UploadOptions): After<RecordActionResponse> => {
-      const feature = uploadFile(options)({})
+      const feature = uploadFile(options)(new AdminJS(), {})
       return feature.actions?.show?.after?.[0] as After<RecordActionResponse>
     }
 
     it('fills record with the path', async () => {
-      const response = { record: { params: {
-        [properties.key]: key,
-      } } }
-      const fillPath = getAfterHook({ provider, properties })
+      const response = {
+        record: {
+          params: {
+            [properties.key]: key,
+          },
+        },
+      }
+
+      const fillPath = getAfterHook({ componentLoader, provider, properties })
 
       const ret = await fillPath(
         response as RecordActionResponse,
@@ -66,11 +72,15 @@ describe('uploadFileFeature', () => {
     it('gets bucket from the record when it is present', async () => {
       const bucket = 'some-other-bucket'
       properties.bucket = 'storedBucketProperty'
-      const response = { record: { params: {
-        [properties.key]: key,
-        [properties.bucket]: bucket,
-      } } }
-      const fillPath = getAfterHook({ provider, properties })
+      const response = {
+        record: {
+          params: {
+            [properties.key]: key,
+            [properties.bucket]: bucket,
+          },
+        },
+      }
+      const fillPath = getAfterHook({ componentLoader, provider, properties })
 
       await fillPath(
         response as RecordActionResponse,
@@ -82,10 +92,15 @@ describe('uploadFileFeature', () => {
     })
 
     it('does nothing when path is not present', async () => {
-      const response = { record: { params: {
-        name: 'some value',
-      } } }
-      const fillPath = getAfterHook({ provider, properties })
+      const response = {
+        record: {
+          params: {
+            name: 'some value',
+          },
+        },
+      }
+
+      const fillPath = getAfterHook({ componentLoader, provider, properties })
 
       const ret = await fillPath(
         response as unknown as RecordActionResponse,
