@@ -5,8 +5,8 @@
  * @load ./index.doc.md
  */
 
-import uploadFileFeature from './features/upload-file/upload-file.feature'
+import uploadFileFeature from './features/upload-file/upload-file.feature.js'
 
 export default uploadFileFeature
-export * from './features/upload-file/types/upload-options.type'
-export * from './features/upload-file/providers'
+export * from './features/upload-file/providers/index.js'
+export * from './features/upload-file/types/upload-options.type.js'
